Modify: model classes their services
Modified JSON related annotations of model classes, replaced some NotBlank annotations in DTO classes with NotNull, and optimized delete methods of services

+++ src/main/java/com/internship/blog/book/BookService.java
@@ -31,6 +31,10 @@ public class BookService {
     }
 
     public String deleteBookById(Integer id) {
+        Book book = bookRepository.findById(id).orElse(null);
+        if (book == null) {
+            return "Book not found with id: " + id;
+        }
         bookRepository.deleteById(id);
         return "Book with id " + id + " was deleted";
     }
